Added new xns for Aerojet and L28 canals
Created 2 new xns groups for Aerojet, one prior to the addition of the
C-111SC plugs/weirs, and one that includes the new plugs and weirs.
Also added 2 new xns groups for L28, one with the historic plugs in the
L28 canal, and one with the new raised plugs in the L28.  The levee has
been moved to the appropriate side of the canal.