FIXED model to match LaTeX (Part 2 of 2)
Many inconsistencies have been introduced between the nwk11 file and the
LaTeX documentation of the operations during the past several commits.
This commit fixes the incorrect operations in the model to match the
correct operations in the LaTeX. Some structure attributes were also
corrected.

At this point, the model and the LaTeX should once again be synchronized.
It would probably be beneficial for someone to double check this, since so
many changes have been made.