INPUTFILES: added infiles for WQ
Added infiles for Ecolab/WQ/AD simulations for M01 and M06. Currently they
will only be used with M01_V10 and greater, because this is on the new
shifted grid. files for M06 will work with either M06 OR M01_V10 because
they share the same grid.
Next task is to get M01_V10 working with AD/WQ/Ecolab...