changed M11_OPERATIONS-IOP to 1-minute intervals
The IOP ops file was previously daily intervals, but there was a problem
when switiching values from day to day. MIKE interpolates between the
integer values, causing the timesteps on switching days to have
non-integer MODE values. This would cause the priority rules to not be
activated for sub-daily timesteps on days where the value changed from
one integer to the other.

The solution employed in the new IOP ops file was to change the
switching interval to be only one minute instead of one day on the days
where the value changed.