Added ICs for all MIKE11 points
WE MAY _NOT_ WANT TO USE THESE EDITS

** Also added Mike11.ini file with important setting:
maximum exceedance factor (WL_EXCEEDANCE_FACTOR) was increased provisionally from 4 to 10000

--> I don't know if we want to keep this setting change. I am not sure
of the implications.

--> I don't know if we really need to set ICs for MIKE11. It adds a
bunch of complexity to the model, and more stuff to maintain when making
changes to the network (ie if you shorten a canal you need to remember
to change the chainage in the hd11 file too).

So, I'vs saved the edits here incase we want to pick this up again.