Updated M11 ICs precision
WE MAY NOT WANT TO KEEP THIS - see commit message before last one

Simplified the M11 BCs by removing the significant digits smaller than a
tenth of a foot. (values are saved in meters, however)

I though it would make the ICs easier to maintain and understand,
assuming we want to use them at all.