Added ICs for MIKESHE
WE MAY _NOT_ WANT TO KEEP THESE EDITS

Added initial stages/depth for the MSHE OL and SZ, based on model
results from 1/1/2000.  Before we started the model with initial depth=0
across the model domain.

The problem is, the model will not run unless you increase the
WL_EXCEEDANCE_FACTOR in file Mike11.ini. I tried to get it to run with
this factor set to 4 and to 10, but it crashed. This commit has it set
to 10000.

I am not sure we want to raise the WL_EXCEEDANCE_FACTOR.