Deleted C24-28 in M06
Deleted Control structures for Culverts 24, 25, 26, 27, and 28. These
had prescribed flows. S343A and S343 B also have prescribed flows in the
M06 model, but not in the M01 model.  The culverts were just deleted in
the M06 model.

This edit still needs to be resolved for M01, AND these culverts now
need to be re-added to the model as passive round culverts (not control
structures)