added more M11 latex files
In order to keep all the edits in the M11 model up-to-date with the documentation, I added in the latex sections describing those.

These sections have a lot of figures that I don't want in git right now, so I added a new NON-GIT directory called 'figs-not-in-git' to refer to those.

I also updated the main.tex and other support files to mirror the current edits in the main report (not in git).